+++ src/Umbraco.Core/Persistence/Repositories/VersionableRepositoryBase.cs
@@ -274,7 +274,7 @@ namespace Umbraco.Core.Persistence.Repositories
         /// <param name="orderDirection">The order direction.</param>
         /// <returns></returns>
         /// <exception cref="System.ArgumentNullException">orderBy</exception>
-        protected IEnumerable<TEntity> GetPagedResultsByQuery<TDto,TContentBase>(IQuery<TEntity> query, int pageIndex, int pageSize, out int totalRecords,
+        protected IEnumerable<TEntity> GetPagedResultsByQuery<TDto, TContentBase>(IQuery<TEntity> query, long pageIndex, int pageSize, out long totalRecords,
             Tuple<string, string> nodeIdSelect,
             Func<Sql, IEnumerable<TEntity>> processQuery,
             string orderBy,
