+++ src/Umbraco.Tests/Persistence/Repositories/MediaRepositoryTest.cs
@@ -317,7 +317,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IMedia>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 0, 1, out totalRecords, "SortOrder", Direction.Ascending);
 
                 // Assert
@@ -338,7 +338,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IMedia>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 1, 1, out totalRecords, "SortOrder", Direction.Ascending);
 
                 // Assert
@@ -359,7 +359,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IMedia>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 0, 2, out totalRecords, "SortOrder", Direction.Ascending);
 
                 // Assert
@@ -380,7 +380,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IMedia>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 0, 1, out totalRecords, "SortOrder", Direction.Descending);
 
                 // Assert
@@ -401,7 +401,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IMedia>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 0, 1, out totalRecords, "Name", Direction.Ascending);
 
                 // Assert
@@ -422,7 +422,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IMedia>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 0, 1, out totalRecords, "SortOrder", Direction.Ascending, "File");
 
                 // Assert
@@ -443,7 +443,7 @@ namespace Umbraco.Tests.Persistence.Repositories
             {
                 // Act
                 var query = Query<IMedia>.Builder.Where(x => x.Level == 2);
-                int totalRecords;
+                long totalRecords;
                 var result = repository.GetPagedResultsByQuery(query, 0, 1, out totalRecords, "SortOrder", Direction.Ascending, "Test");
 
                 // Assert
